refacto(Inventory): move files
