+++ src/app/inventory/redux/reducers/entities/inventories/getters/get-busy-storage.js
