+++ src/app/inventory/redux/reducers/entities/inventories/getters/get-has-items-form-inventory.js
@@ -1,6 +1,6 @@
 import { isEmpty } from "ramda";
 
-import getItemsFromCurrentInventory from "../../inventories/getters/get-items-from-current-inventory";
+import getItemsFromCurrentInventory from "./get-items-from-current-inventory";
 
 export default state => {
   const items = getItemsFromCurrentInventory(state);
