+++ src/app/inventory/redux/reducers/entities/inventories/getters/get-item-completed-from-inventory-by-id.js
@@ -1,6 +1,6 @@
 import { pathOr } from "ramda";
 
-import getItemFromCurrentInventoryById from "../../inventories/getters/get-item-from-current-inventory-by-id";
+import getItemFromCurrentInventoryById from "./get-item-from-current-inventory-by-id";
 import selectItemById from "../../items/selectors/select-items-by-id";
 
 export default (state, id) => {
